Move initializer data function
