+++ packages/plugin-hardhat/src/deploy-proxy.ts
@@ -12,8 +12,8 @@ import {
   DeployTransaction,
   DeployKindUnsupported,
   deployProxyImpl,
+  getInitializerData,
 } from './utils';
-import { Interface } from '@ethersproject/abi';
 
 export interface DeployFunction {
   (ImplFactory: ContractFactory, args?: unknown[], opts?: DeployOptions): Promise<Contract>;
@@ -78,29 +78,4 @@ export function makeDeployProxy(hre: HardhatRuntimeEnvironment): DeployFunction 
     inst.deployTransaction = proxyDeployment.deployTransaction;
     return inst;
   };
-}
-
-export function getInitializerData(
-  contractInterface: Interface,
-  args: unknown[],
-  initializer?: string | false,
-): string {
-  if (initializer === false) {
-    return '0x';
-  }
-
-  const allowNoInitialization = initializer === undefined && args.length === 0;
-  initializer = initializer ?? 'initialize';
-
-  try {
-    const fragment = contractInterface.getFunction(initializer);
-    return contractInterface.encodeFunctionData(fragment, args);
-  } catch (e: unknown) {
-    if (e instanceof Error) {
-      if (allowNoInitialization && e.message.includes('no matching function')) {
-        return '0x';
-      }
-    }
-    throw e;
-  }
 }
