+++ packages/plugin-hardhat/src/utils/initializer-data.ts
@@ -1,0 +1,26 @@
+import { Interface } from '@ethersproject/abi';
+
+export function getInitializerData(
+  contractInterface: Interface,
+  args: unknown[],
+  initializer?: string | false,
+): string {
+  if (initializer === false) {
+    return '0x';
+  }
+
+  const allowNoInitialization = initializer === undefined && args.length === 0;
+  initializer = initializer ?? 'initialize';
+
+  try {
+    const fragment = contractInterface.getFunction(initializer);
+    return contractInterface.encodeFunctionData(fragment, args);
+  } catch (e: unknown) {
+    if (e instanceof Error) {
+      if (allowNoInitialization && e.message.includes('no matching function')) {
+        return '0x';
+      }
+    }
+    throw e;
+  }
+}
